refactor: split monolithic utils.py into focused modules
- if_llm/image_utils.py — image processing, tensor/PIL/base64 conversions, mask handling
- if_llm/text_utils.py — clean_text text cleaning utility
- if_llm/model_utils.py — API key management, model listing, HuggingFace helpers
- if_llm/gemini2_utils.py — Gemini 2.0 image processing and client utilities
- if_llm/settings_utils.py — YAML dumping, combo settings management
- utils.py — backward-compatible re-export module
- Updated imports in send_request.py, ListModelsNode.py, IFLLMNode.py
